Refactor for the FastEventManager
